Movement methods renamed, forces now in PointMass class
